perf: Split config handler
