+++ src/request-handler.ts
@@ -3,81 +3,29 @@ import type {
   DefaultResponse,
   RequestHandlerConfig,
   RequestConfig,
-  Method,
   RetryOptions,
   FetchResponse,
   RequestHandlerReturnType,
   CreatedCustomFetcherInstance,
-  FetcherConfig,
-  FetcherInstance,
-  Logger,
-  HeadersObject,
 } from './types/request-handler';
 import type {
-  BodyPayload,
   DefaultParams,
   DefaultPayload,
   DefaultUrlParams,
-  QueryParams,
 } from './types/api-handler';
 import { applyInterceptor } from './interceptor-manager';
 import { ResponseError } from './errors/response-error';
 import {
-  appendQueryParams,
-  isJSONSerializable,
-  replaceUrlPathParams,
   delayInvocation,
   flattenData,
   processHeaders,
-  isSearchParams,
   sanitizeObject,
 } from './utils';
-import { addRequest, removeRequest } from './queue-manager';
-import {
-  ABORT_ERROR,
-  APPLICATION_JSON,
-  CANCELLED_ERROR,
-  CHARSET_UTF_8,
-  CONTENT_TYPE,
-  FUNCTION,
-  GET,
-  HEAD,
-  OBJECT,
-  STRING,
-  UNDEFINED,
-} from './constants';
+import { queueRequest, removeRequestFromQueue } from './queue-manager';
+import { ABORT_ERROR, CANCELLED_ERROR, OBJECT } from './constants';
 import { parseResponseData } from './response-parser';
 import { generateCacheKey, getCache, setCache } from './cache-manager';
-
-const defaultConfig: RequestHandlerConfig = {
-  method: GET,
-  strategy: 'reject',
-  timeout: 30000,
-  dedupeTime: 1000,
-  defaultResponse: null,
-  headers: {
-    Accept: APPLICATION_JSON + ', text/plain, */*',
-    'Accept-Encoding': 'gzip, deflate, br',
-  },
-  retry: {
-    delay: 1000,
-    maxDelay: 30000,
-    resetTimeout: true,
-    backoff: 1.5,
-
-    // https://developer.mozilla.org/en-US/docs/Web/HTTP/Status
-    retryOn: [
-      408, // Request Timeout
-      409, // Conflict
-      425, // Too Early
-      429, // Too Many Requests
-      500, // Internal Server Error
-      502, // Bad Gateway
-      503, // Service Unavailable
-      504, // Gateway Timeout
-    ],
-  },
-};
+import { buildConfig, defaultConfig, mergeConfig } from './config-handler';
 
 /**
  * Create Request Handler
@@ -94,52 +42,13 @@ export function createRequestHandler(
     ...sanitizedConfig,
   };
 
-  /**
-   * Merges the specified property from the base configuration and the new configuration into the target configuration.
-   *
-   * @param {K} property - The property key to merge from the base and new configurations. Must be a key of RequestHandlerConfig.
-   * @param {RequestHandlerConfig} targetConfig - The configuration object that will receive the merged properties.
-   * @param {RequestHandlerConfig} baseConfig - The base configuration object that provides default values.
-   * @param {RequestHandlerConfig} newConfig - The new configuration object that contains user-specific settings to merge.
-   */
-  const mergeConfig = <K extends keyof RequestHandlerConfig>(
-    property: K,
-    targetConfig: RequestHandlerConfig,
-    baseConfig: RequestHandlerConfig,
-    newConfig: RequestHandlerConfig,
-  ) => {
-    if (newConfig[property]) {
-      targetConfig[property] = {
-        ...baseConfig[property],
-        ...newConfig[property],
-      };
-    }
-  };
-
   mergeConfig('retry', handlerConfig, defaultConfig, sanitizedConfig);
   mergeConfig('headers', handlerConfig, defaultConfig, sanitizedConfig);
 
   /**
-   * Gets a configuration value from `reqConfig`, defaulting to `handlerConfig` if not present.
-   *
-   * @param {RequestConfig} reqConfig - Request configuration object.
-   * @param {keyof RequestConfig} name - Key of the configuration value.
-   * @returns {T} - The configuration value.
-   */
-  const getConfig = <T = unknown>(
-    reqConfig: RequestConfig,
-    name: keyof RequestConfig,
-  ): T => {
-    return typeof reqConfig[name] !== UNDEFINED
-      ? reqConfig[name]
-      : handlerConfig[name];
-  };
-
-  /**
    * Immediately create instance of custom fetcher if it is defined
    */
-  const customFetcher = getConfig<FetcherInstance>(sanitizedConfig, 'fetcher');
-  const requestInstance = customFetcher?.create(handlerConfig) || null;
+  const requestInstance = sanitizedConfig.fetcher?.create?.(handlerConfig);
 
   /**
    * Get Provider Instance
@@ -147,138 +56,7 @@ export function createRequestHandler(
    * @returns {CreatedCustomFetcherInstance | null} Provider's instance
    */
   const getInstance = (): CreatedCustomFetcherInstance | null => {
-    return requestInstance;
-  };
-
-  /**
-   * Logs messages or errors using the configured logger's `warn` method.
-   *
-   * @param {RequestConfig} reqConfig - Request config passed when making the request
-   * @param {...(string | ResponseError<any>)} args - Messages or errors to log.
-   */
-  const logger = (
-    reqConfig: RequestConfig,
-    ...args: (string | ResponseError<any>)[]
-  ): void => {
-    const logger = getConfig<Logger>(reqConfig, 'logger');
-
-    if (logger?.warn) {
-      logger.warn(...args);
-    }
-  };
-
-  /**
-   * Ensures the `Content-Type` header is set to `application/json; charset=utf-8`
-   * if it is not already present and the request method and body meet specific conditions.
-   *
-   * @param headers - The headers object to modify. Can be an instance of `Headers`
-   *                  or a plain object conforming to `HeadersInit`.
-   * @param method - The HTTP method of the request (e.g., 'PUT', 'DELETE', etc.).
-   * @param body - The optional body of the request. If no body is provided and the
-   *               method is 'PUT' or 'DELETE', the function exits without modifying headers.
-   */
-  const setContentTypeIfNeeded = (
-    headers: HeadersInit,
-    method: string,
-    body?: unknown,
-  ): void => {
-    if (!body && ['PUT', 'DELETE'].includes(method)) {
-      return;
-    }
-
-    const contentTypeValue = APPLICATION_JSON + ';' + CHARSET_UTF_8;
-
-    if (headers instanceof Headers) {
-      if (!headers.has(CONTENT_TYPE)) {
-        headers.set(CONTENT_TYPE, contentTypeValue);
-      }
-    } else if (
-      typeof headers === OBJECT &&
-      !Array.isArray(headers) &&
-      !headers[CONTENT_TYPE]
-    ) {
-      headers[CONTENT_TYPE] = contentTypeValue;
-    }
-  };
-
-  /**
-   * Build request configuration
-   *
-   * @param {string} url - Request url
-   * @param {RequestConfig} requestConfig - Request config passed when making the request
-   * @returns {RequestConfig} - Provider's instance
-   */
-  const buildConfig = (
-    url: string,
-    requestConfig: RequestConfig,
-  ): FetcherConfig => {
-    const method = getConfig<string>(
-      requestConfig,
-      'method',
-    ).toUpperCase() as Method;
-    const isGetAlikeMethod = method === GET || method === HEAD;
-
-    const dynamicUrl = replaceUrlPathParams(
-      url,
-      getConfig(requestConfig, 'urlPathParams'),
-    );
-
-    // The explicitly passed query params
-    const explicitParams = getConfig<QueryParams>(requestConfig, 'params');
-
-    // Final body data
-    let body: RequestConfig['data'];
-
-    // Only applicable for request methods 'PUT', 'POST', 'DELETE', and 'PATCH'
-    if (!isGetAlikeMethod) {
-      // The explicitly passed "body" or "data"
-      const explicitBodyData: BodyPayload =
-        getConfig(requestConfig, 'body') || getConfig(requestConfig, 'data');
-
-      body = explicitBodyData;
-    }
-
-    const headers = getConfig<HeadersObject>(requestConfig, 'headers');
-
-    setContentTypeIfNeeded(headers, method, body);
-
-    // Native fetch compatible settings
-    const isWithCredentials = getConfig<boolean>(
-      requestConfig,
-      'withCredentials',
-    );
-
-    const credentials = isWithCredentials
-      ? 'include'
-      : getConfig<RequestCredentials>(requestConfig, 'credentials');
-
-    const urlPath = explicitParams
-      ? appendQueryParams(dynamicUrl, explicitParams)
-      : dynamicUrl;
-    const isFullUrl = urlPath.includes('://');
-    const baseURL = isFullUrl
-      ? ''
-      : getConfig<string>(requestConfig, 'baseURL') ||
-        getConfig<string>(requestConfig, 'apiUrl');
-
-    // Automatically stringify request body, if possible and when not dealing with strings
-    if (
-      body &&
-      typeof body !== STRING &&
-      !isSearchParams(body) &&
-      isJSONSerializable(body)
-    ) {
-      body = JSON.stringify(body);
-    }
-
-    return {
-      ...requestConfig,
-      credentials,
-      body,
-      method,
-      headers,
-      url: baseURL + urlPath,
-    };
+    return requestInstance || null;
   };
 
   /**
@@ -336,22 +114,21 @@ export function createRequestHandler(
       RequestBody
     >,
   ): Promise<any> => {
-    const _isRequestCancelled = isRequestCancelled(error as ResponseError);
-    const errorHandlingStrategy = getConfig<string>(requestConfig, 'strategy');
-    const rejectCancelled = getConfig<boolean>(
-      requestConfig,
-      'rejectCancelled',
-    );
-
     // By default cancelled requests aren't rejected (softFail strategy)
-    if (!(_isRequestCancelled && !rejectCancelled)) {
-      // Hang the promise
-      if (errorHandlingStrategy === 'silent') {
-        await new Promise(() => null);
-      }
+    // Only handle the error if the request was not cancelled,
+    // or if it was cancelled and rejectCancelled is true
+    const isCancelled = isRequestCancelled(error as ResponseError);
+    const shouldHandleError = !isCancelled || requestConfig.rejectCancelled;
+
+    if (shouldHandleError) {
+      const errorHandlingStrategy = requestConfig.strategy;
+
       // Reject the promise
-      else if (errorHandlingStrategy === 'reject') {
+      if (errorHandlingStrategy === 'reject') {
         return Promise.reject(error);
+      } // Hang the promise
+      else if (errorHandlingStrategy === 'silent') {
+        await new Promise(() => null);
       }
     }
 
@@ -360,16 +137,6 @@ export function createRequestHandler(
       requestConfig,
       error,
     );
-  };
-
-  /**
-   * Output error response depending on chosen strategy
-   *
-   * @param {ResponseError} error               Error instance
-   * @returns {boolean}                        True if request is aborted
-   */
-  const isRequestCancelled = (error: ResponseError): boolean => {
-    return error.name === ABORT_ERROR || error.name === CANCELLED_ERROR;
   };
 
   /**
@@ -461,7 +228,7 @@ export function createRequestHandler(
     while (attempt <= _retries) {
       try {
         // Add the request to the queue. Make sure to handle deduplication, cancellation, timeouts in accordance to retry settings
-        const controller = await addRequest(
+        const controller = await queueRequest(
           fetcherConfig,
           timeout,
           dedupeTime,
@@ -483,18 +250,12 @@ export function createRequestHandler(
         // Global interceptors
         await applyInterceptor(requestConfig, handlerConfig?.onRequest);
 
-        if (
-          customFetcher !== null &&
-          requestInstance !== null &&
-          typeof requestInstance.request === FUNCTION
-        ) {
-          response = await requestInstance.request(requestConfig);
-        } else {
-          response = (await fetch(
-            requestConfig.url as string,
-            requestConfig as RequestInit,
-          )) as unknown as FetchResponse<ResponseData, RequestBody>;
-        }
+        response = requestInstance?.request
+          ? await requestInstance.request(requestConfig)
+          : ((await fetch(
+              requestConfig.url as string,
+              requestConfig as RequestInit,
+            )) as unknown as FetchResponse<ResponseData, RequestBody>);
 
         // Add more information to response object
         if (response instanceof Response) {
@@ -522,7 +283,7 @@ export function createRequestHandler(
         // Global interceptors
         await applyInterceptor(response, handlerConfig?.onResponse);
 
-        removeRequest(fetcherConfig);
+        removeRequestFromQueue(fetcherConfig);
 
         if (
           shouldRetry &&
@@ -553,7 +314,7 @@ export function createRequestHandler(
           // Restart the main retry loop
           pollingAttempt++;
 
-          logger(requestConfig, 'Polling attempt ' + pollingAttempt + '...');
+          logger(mergedConfig, 'Polling attempt ' + pollingAttempt + '...');
 
           await delayInvocation(pollingInterval);
 
@@ -606,7 +367,7 @@ export function createRequestHandler(
             RequestBody
           >(error, fetcherConfig);
 
-          removeRequest(fetcherConfig);
+          removeRequestFromQueue(fetcherConfig);
 
           return outputErrorResponse<
             ResponseData,
@@ -663,7 +424,7 @@ export function createRequestHandler(
       RequestBody
     > | null = null,
   ): FetchResponse<ResponseData, RequestBody> => {
-    const defaultResponse = getConfig<any>(requestConfig, 'defaultResponse');
+    const defaultResponse = requestConfig.defaultResponse ?? null;
 
     // This may happen when request is cancelled.
     if (!response) {
@@ -688,13 +449,7 @@ export function createRequestHandler(
       data = defaultResponse;
     }
 
-    // Return flattened response immediately
-    const flattenResponse = getConfig<boolean>(
-      requestConfig,
-      'flattenResponse',
-    );
-
-    if (flattenResponse) {
+    if (requestConfig.flattenResponse) {
       response.data = flattenData(data);
     }
 
@@ -733,8 +488,34 @@ export function createRequestHandler(
 
   return {
     getInstance,
-    buildConfig,
     config: handlerConfig,
     request,
   };
 }
+
+/**
+ * Output error response depending on chosen strategy
+ *
+ * @param {ResponseError} error               Error instance
+ * @returns {boolean}                        True if request is aborted
+ */
+const isRequestCancelled = (error: ResponseError): boolean => {
+  return error.name === ABORT_ERROR || error.name === CANCELLED_ERROR;
+};
+
+/**
+ * Logs messages or errors using the configured logger's `warn` method.
+ *
+ * @param {RequestConfig} reqConfig - Request config passed when making the request
+ * @param {...(string | ResponseError<any>)} args - Messages or errors to log.
+ */
+const logger = (
+  reqConfig: RequestConfig,
+  ...args: (string | ResponseError<any>)[]
+): void => {
+  const logger = reqConfig.logger;
+
+  if (logger?.warn) {
+    logger.warn(...args);
+  }
+};
