+++ test/queue-manager.spec.ts
@@ -1,4 +1,8 @@
-import { addRequest, removeRequest, getController } from '../src/queue-manager';
+import {
+  queueRequest,
+  removeRequestFromQueue,
+  getController,
+} from '../src/queue-manager';
 import type { RequestConfig } from '../src/types';
 
 const createConfig = (url: string): RequestConfig => ({
@@ -14,15 +18,15 @@ describe('Request Queue Manager', () => {
   });
 
   it('should add and retrieve a request correctly', async () => {
-    const controller = await addRequest(config, 1000);
+    const controller = await queueRequest(config, 1000);
     const retrievedController = await getController(config);
 
     expect(retrievedController).toBe(controller);
   });
 
   it('should remove a request from the queue', async () => {
-    await addRequest(config, 1000);
-    await removeRequest(config);
+    await queueRequest(config, 1000);
+    await removeRequestFromQueue(config);
     const retrievedController = await getController(config);
 
     expect(retrievedController).toBeUndefined();
@@ -30,7 +34,7 @@ describe('Request Queue Manager', () => {
 
   it('should handle removing a non-existent request', async () => {
     await expect(
-      removeRequest(createConfig('https://example.com')),
+      removeRequestFromQueue(createConfig('https://example.com')),
     ).resolves.not.toThrow();
   });
 
@@ -40,8 +44,8 @@ describe('Request Queue Manager', () => {
 
     // Start concurrent requests
     const [controller1, controller2] = await Promise.all([
-      addRequest(config1, 1000),
-      addRequest(config2, 1000),
+      queueRequest(config1, 1000),
+      queueRequest(config2, 1000),
     ]);
 
     // Ensure controllers are retrieved correctly
@@ -54,8 +58,8 @@ describe('Request Queue Manager', () => {
     expect(retrievedController2).toBe(controller2);
 
     // Cleanup
-    await removeRequest(config1);
-    await removeRequest(config2);
+    await removeRequestFromQueue(config1);
+    await removeRequestFromQueue(config2);
   });
 
   it('should handle concurrent requests with different configurations separately', async () => {
@@ -64,8 +68,8 @@ describe('Request Queue Manager', () => {
 
     // Add two concurrent requests with different configurations
     const [controller1, controller2] = await Promise.all([
-      addRequest(config1, 2000),
-      addRequest(config2, 2000),
+      queueRequest(config1, 2000),
+      queueRequest(config2, 2000),
     ]);
 
     jest.advanceTimersByTime(2000);
@@ -80,7 +84,7 @@ describe('Request Queue Manager', () => {
     const timeout = 1000;
 
     // Add a request
-    await addRequest(config, timeout);
+    await queueRequest(config, timeout);
 
     // Advance timers to simulate timeout
     jest.advanceTimersByTime(timeout);
@@ -90,17 +94,17 @@ describe('Request Queue Manager', () => {
     expect(controller).toBeUndefined(); // Ensure the request was removed
 
     // Ensure request removal
-    await removeRequest(config);
+    await removeRequestFromQueue(config);
   });
 
   it('should queue multiple operations on the same request config correctly', async () => {
     const config = createConfig('https://example.com');
 
     // Simulate a long-running first request
-    const firstRequestPromise = addRequest(config, 2000);
+    const firstRequestPromise = queueRequest(config, 2000);
 
     // Attempt to add a second request that should be queued
-    const secondRequestPromise = addRequest(config, 2000);
+    const secondRequestPromise = queueRequest(config, 2000);
 
     // Advance timers to simulate part of the delay for the first request
     jest.advanceTimersByTime(500);
@@ -125,7 +129,7 @@ describe('Request Queue Manager', () => {
 
   it('should clear timeout and abort request on removal', async () => {
     const configWithTimeout = createConfig('https://example.com');
-    await addRequest(configWithTimeout, 1000);
+    await queueRequest(configWithTimeout, 1000);
 
     jest.advanceTimersByTime(1500);
 
@@ -136,10 +140,10 @@ describe('Request Queue Manager', () => {
 
   it('should deduplicate same requests within dedupeTime', async () => {
     // Add a request to the queue
-    const controller1 = await addRequest(mockConfig, 2000, 1000);
+    const controller1 = await queueRequest(mockConfig, 2000, 1000);
 
     // Trigger the same request within the dedupeTime
-    const controller2 = await addRequest(mockConfig, 2000, 1000);
+    const controller2 = await queueRequest(mockConfig, 2000, 1000);
 
     jest.advanceTimersByTime(500);
 
@@ -149,10 +153,10 @@ describe('Request Queue Manager', () => {
 
   it('should not deduplicate requests if dedupeTime is 0', async () => {
     // Add the first request
-    const controller1 = await addRequest(mockConfig, 1000, 0);
+    const controller1 = await queueRequest(mockConfig, 1000, 0);
 
     // Add the same request with dedupeTime 0
-    const controller2 = await addRequest(mockConfig, 1000, 0);
+    const controller2 = await queueRequest(mockConfig, 1000, 0);
 
     // Fast-forward time (though not necessary for 0 dedupeTime)
     jest.advanceTimersByTime(1000);
@@ -162,12 +166,12 @@ describe('Request Queue Manager', () => {
   });
 
   it('should not abort the request when timeout is disabled', async () => {
-    const controller = await addRequest(mockConfig, 0, 0, false, false);
+    const controller = await queueRequest(mockConfig, 0, 0, false, false);
 
     jest.advanceTimersByTime(1000);
 
     expect(controller.signal.aborted).toBe(false);
-    await removeRequest(mockConfig, null);
+    await removeRequestFromQueue(mockConfig, null);
   });
 
   it('should handle multiple distinct requests separately', async () => {
@@ -181,8 +185,8 @@ describe('Request Queue Manager', () => {
     };
 
     // Add two distinct requests
-    const controllerA = await addRequest(configA, 1000, 1000);
-    const controllerB = await addRequest(configB, 1000, 1000);
+    const controllerA = await queueRequest(configA, 1000, 1000);
+    const controllerB = await queueRequest(configB, 1000, 1000);
 
     jest.advanceTimersByTime(1000);
 
@@ -194,8 +198,8 @@ describe('Request Queue Manager', () => {
     const timeout = 1000;
 
     // Add multiple requests
-    const controller1 = await addRequest(config, timeout, 1000, true);
-    const controller2 = await addRequest(config, timeout, 1000, true);
+    const controller1 = await queueRequest(config, timeout, 1000, true);
+    const controller2 = await queueRequest(config, timeout, 1000, true);
 
     // Advance timers to simulate timeout
     jest.advanceTimersByTime(timeout + 500);
@@ -206,8 +210,8 @@ describe('Request Queue Manager', () => {
 
   it('should handle requests with the same configuration but different options correctly', async () => {
     const config = createConfig('https://example.com');
-    const controller1 = await addRequest(config, 2000, 1000, true); // Cancellable
-    const controller2 = await addRequest(config, 2000, 1000, false); // Not cancellable
+    const controller1 = await queueRequest(config, 2000, 1000, true); // Cancellable
+    const controller2 = await queueRequest(config, 2000, 1000, false); // Not cancellable
 
     // Advance timers to simulate request handling
     jest.advanceTimersByTime(1500);
@@ -223,8 +227,8 @@ describe('Request Queue Manager', () => {
     config2.method = 'POST'; // Change method to simulate different config
 
     // Add requests with different configurations
-    const controller1 = await addRequest(config1, 2000, 1000);
-    const controller2 = await addRequest(config2, 2000, 1000);
+    const controller1 = await queueRequest(config1, 2000, 1000);
+    const controller2 = await queueRequest(config2, 2000, 1000);
 
     // Advance timers to simulate request handling
     jest.advanceTimersByTime(1500);
@@ -235,9 +239,9 @@ describe('Request Queue Manager', () => {
   });
 
   it('should cancel all previous requests if they are cancellable and deduplication time is not yet passed', async () => {
-    const controller1 = await addRequest(mockConfig, 2000, 1000, true);
-    const controller2 = await addRequest(mockConfig, 2000, 1000, true);
-    const controller3 = await addRequest(mockConfig, 2000, 1000, true);
+    const controller1 = await queueRequest(mockConfig, 2000, 1000, true);
+    const controller2 = await queueRequest(mockConfig, 2000, 1000, true);
+    const controller3 = await queueRequest(mockConfig, 2000, 1000, true);
 
     jest.advanceTimersByTime(500);
 
@@ -249,13 +253,13 @@ describe('Request Queue Manager', () => {
     expect(controller3).toBeInstanceOf(AbortController);
 
     jest.spyOn(Date, 'now').mockRestore();
-    await removeRequest(mockConfig);
+    await removeRequestFromQueue(mockConfig);
   });
 
   it('should cancel all previous requests if they are cancellable and deduplication time is passed', async () => {
-    const controller1 = await addRequest(mockConfig, 2000, 1000, true);
-    const controller2 = await addRequest(mockConfig, 2000, 1000, true);
-    const controller3 = await addRequest(mockConfig, 2000, 1000, true);
+    const controller1 = await queueRequest(mockConfig, 2000, 1000, true);
+    const controller2 = await queueRequest(mockConfig, 2000, 1000, true);
+    const controller3 = await queueRequest(mockConfig, 2000, 1000, true);
 
     jest.advanceTimersByTime(1500);
 
@@ -267,13 +271,13 @@ describe('Request Queue Manager', () => {
     expect(controller3).toBeInstanceOf(AbortController);
 
     jest.spyOn(Date, 'now').mockRestore();
-    await removeRequest(mockConfig);
+    await removeRequestFromQueue(mockConfig);
   });
 
   it('should cancel all requests if they are cancellable and timeout is passed', async () => {
-    const controller1 = await addRequest(mockConfig, 2000, 1000, true);
-    const controller2 = await addRequest(mockConfig, 2000, 1000, true);
-    const controller3 = await addRequest(mockConfig, 2000, 1000, true);
+    const controller1 = await queueRequest(mockConfig, 2000, 1000, true);
+    const controller2 = await queueRequest(mockConfig, 2000, 1000, true);
+    const controller3 = await queueRequest(mockConfig, 2000, 1000, true);
 
     jest.advanceTimersByTime(2500);
 
@@ -285,13 +289,13 @@ describe('Request Queue Manager', () => {
     expect(controller3).toBeInstanceOf(AbortController);
 
     jest.spyOn(Date, 'now').mockRestore();
-    await removeRequest(mockConfig);
+    await removeRequestFromQueue(mockConfig);
   });
 
   it('should not cancel any request if not cancellable and deduplication time is not yet passed', async () => {
-    const controller1 = await addRequest(mockConfig, 2000, 1000, false);
-    const controller2 = await addRequest(mockConfig, 2000, 1000, false);
-    const controller3 = await addRequest(mockConfig, 2000, 1000, false);
+    const controller1 = await queueRequest(mockConfig, 2000, 1000, false);
+    const controller2 = await queueRequest(mockConfig, 2000, 1000, false);
+    const controller3 = await queueRequest(mockConfig, 2000, 1000, false);
 
     jest.advanceTimersByTime(500);
 
@@ -303,15 +307,15 @@ describe('Request Queue Manager', () => {
     expect(controller3).toBeInstanceOf(AbortController);
 
     jest.spyOn(Date, 'now').mockRestore();
-    await removeRequest(mockConfig);
+    await removeRequestFromQueue(mockConfig);
   });
 
   it('should not cancel any request if not cancellable and deduplication time is passed for each request', async () => {
-    const controller1 = await addRequest(mockConfig, 20000, 1000, false);
+    const controller1 = await queueRequest(mockConfig, 20000, 1000, false);
     jest.advanceTimersByTime(1500);
-    const controller2 = await addRequest(mockConfig, 20000, 1000, false);
+    const controller2 = await queueRequest(mockConfig, 20000, 1000, false);
     jest.advanceTimersByTime(1500);
-    const controller3 = await addRequest(mockConfig, 20000, 1000, false);
+    const controller3 = await queueRequest(mockConfig, 20000, 1000, false);
     jest.advanceTimersByTime(1500);
 
     expect(controller1).not.toBe(controller3);
@@ -323,15 +327,15 @@ describe('Request Queue Manager', () => {
     expect(controller3).toBeInstanceOf(AbortController);
 
     jest.spyOn(Date, 'now').mockRestore();
-    await removeRequest(mockConfig);
+    await removeRequestFromQueue(mockConfig);
   });
 
   it('should not cancel the previous requests if they are not cancellable', async () => {
-    const controller1 = await addRequest(mockConfig, 2000, 1000, false);
+    const controller1 = await queueRequest(mockConfig, 2000, 1000, false);
 
     jest.advanceTimersByTime(1500);
 
-    const controller2 = await addRequest(mockConfig, 2000, 1000, true);
+    const controller2 = await queueRequest(mockConfig, 2000, 1000, true);
 
     expect(controller1).not.toBe(controller2);
     expect(controller1.signal.aborted).toBe(false);
@@ -339,6 +343,6 @@ describe('Request Queue Manager', () => {
     expect(controller2).toBeInstanceOf(AbortController);
 
     jest.spyOn(Date, 'now').mockRestore();
-    await removeRequest(mockConfig);
+    await removeRequestFromQueue(mockConfig);
   });
 });
